Create krit_gates.lua
Combining the toggles and ranges tutorials from this page: https://monome.org/docs/norns/grid-recipes/
to recreate the 'gates' page from Kria.

ATTN line 66: seems to cancel the  redraw clock. I think the issue is that I need to differentiate the clocks from one another in the code, but I'm uncertain of how to do that and have given it Plenty Of Time.